backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "ortools>=9.14.6206",
    "scipy>=1.16.0",
]
